Never evaluate unpaginated queries.
Whenever we use `if queryset`, Django actually evaluates the QuerySet
and queries the database. There were a few points in the code where we
did that for the full, non-paginated queryset on every page load.

With this patch, we avoid that and, where we must check whether there
are search results, we evaluate the paginated queryset instead, which
has suitable OFFSET / LIMIT clauses in the underlying SQL and is
therefore cheaper.